refactor: moves test files out of test folder
