+++ src/app/core/services/http/http-rundown.service.spec.ts
@@ -1,8 +1,8 @@
-import { HttpRundownService } from '../http-rundown.service'
-import { HttpErrorService } from '../../../../shared/services/http/http-error.service'
+import { HttpRundownService } from './http-rundown.service'
+import { HttpErrorService } from '../../../shared/services/http/http-error.service'
 import { instance, mock } from '@typestrong/ts-mockito'
 import { HttpClient } from '@angular/common/http'
-import { EntityParser } from '../../../abstractions/entity-parser.service'
+import { EntityParser } from '../../abstractions/entity-parser.service'
 
 describe('HttpRundownService', () => {
   it('should be created', () => {
